introduced model module and moved classes for complex number expression
