+++ app/build.gradle
@@ -26,6 +26,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(":model")
 
     implementation "com.google.android.exoplayer:exoplayer-core:$versions.exo_player"
     implementation "com.google.android.exoplayer:exoplayer-ui:$versions.exo_player"
@@ -51,7 +52,7 @@ dependencies {
     androidTestImplementation "androidx.test.ext:junit:$versions.test_ext_junit"
     androidTestImplementation "androidx.test.ext:truth:$versions.test_ext_truth"
     androidTestImplementation "com.google.truth:truth:$versions.test_truth"
-    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
+    androidTestImplementation("androidx.test.espresso:espresso-core:$versions.espresso", {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 }
