+++ app/src/main/java/com/freshdigitable/ohlplayer/StereoHRTFConvoTask.java
@@ -6,6 +6,7 @@ import androidx.annotation.NonNull;
 import com.freshdigitable.ohlplayer.ImpulseResponse.CHANNEL;
 import com.freshdigitable.ohlplayer.ImpulseResponse.DIRECTION;
 import com.freshdigitable.ohlplayer.ImpulseResponse.SAMPLING_FREQ;
+import com.freshdigitable.ohlplayer.model.ComplexArray;
 
 import java.io.IOException;
 import java.util.Arrays;
