+++ model/build.gradle
@@ -1,0 +1,46 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+android {
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
+
+
+    defaultConfig {
+        minSdkVersion 15
+        targetSdkVersion 29
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles 'consumer-rules.pro'
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation 'androidx.core:core-ktx:1.0.2'
+
+    implementation "androidx.appcompat:appcompat:$versions.appcompat"
+
+    testImplementation "junit:junit:$versions.junit"
+    testImplementation "androidx.test:core:$versions.test_core"
+    testImplementation "androidx.test:runner:$versions.test_runner"
+    testImplementation "androidx.test:rules:$versions.test_rule"
+    testImplementation "androidx.test.ext:junit:$versions.test_ext_junit"
+    testImplementation "androidx.test.ext:truth:$versions.test_ext_truth"
+    testImplementation "com.google.truth:truth:$versions.test_truth"
+
+    androidTestImplementation "androidx.test.ext:junit:$versions.test_ext_junit"
+    androidTestImplementation "androidx.test.espresso:espresso-core:$versions.espresso"
+}
