+++ model/src/test/java/com/freshdigitable/ohlplayer/model/CalcUtilTest.java
@@ -1,4 +1,4 @@
-package com.freshdigitable.ohlplayer;
+package com.freshdigitable.ohlplayer.model;
 
 import org.junit.Before;
 import org.junit.Test;
